Added markdown support.
Replaced gulp-html-replace with swig variables following the gulp-data model.

Added support for markdown within the "markdown" folder.
Markdown files follow the jekyll model with front matter getting pulled out,
the contents compiled with marked, and then the contents wrapped in a
little bit of swig so the "layout" variable selects a swig template.

i.e. {% extends $layout %}

Also refactored the git revision settings into a pipeline.

Next step is to add support for collecting posts
and defining slug urls in some "way" so this could be a drop-in
replacement for jekyll.